# util/deephash: expand fast-path capabilities